YaMetric/requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.2.2
dnspython==2.6.1
filelock==3.14.0
frozenlist==1.4.1
grpcio==1.64.0
grpcio-tools==1.64.0
idna==2.10
magic-filter==1.0.12
multidict==6.0.5
protobuf==5.27.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
setuptools==70.0.0
typing_extensions==4.12.0
yarl==1.9.4
email-validator==2.2.0
requests==2.28.2